Support boolean type in Argument (#15)
* Add support for boolean type

* Add Tests
